Fix byobu and vim ctrl keys
cp should not fail

+++ makesymlinks.sh
@@ -14,7 +14,7 @@ pushd $dir >/dev/null
 for file in ${files}; do
     echo "Creating symlink from ${file} to ~/${file}"
 
-    cp -L ~/${file} ${olddir}
+    [[ -e ~/${file} ]] && cp -L ~/${file} ${olddir}
     ln -f -s ${dir}/${file} ~/${file}
 done
